拆分 MakeTree 相关函数为单独的trait ClosureTable 增加直接父子级查询
